Fix the fork recovery depth metric recording
The change computes the depth as the number of blocks applied since the base block on the fork
being abandoned, at the last moment before switching to the new fork.

The previous approach did not account for the blocks arrived to the abandoned fork after
the recovery process was initiated. Besides, it was too sensitive to the block application time.